requirements.txt
apturl==0.5.2
bcrypt==3.2.0
blinker==1.6.2
Brlapi==0.8.3
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
dnspython==2.3.0
docopt==0.6.2
dominate==2.7.0
duplicity==0.8.21
fasteners==0.14.1
Flask==2.2.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
future==0.18.2
greenlet==2.0.2
gunicorn==20.1.0
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.2
mongo==0.2.0
monotonic==1.6
more-itertools==8.10.0
natsort==8.0.2
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pipreqs==0.4.13
protobuf==3.12.4
ptyprocess==0.7.0
py3dns==3.2.1
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.3.3
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.21.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.28.2
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==2.0.4
stripe==5.4.0
systemd-python==234
typing_extensions==4.5.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.14
usb-creator==0.3.7
validate-email==1.3
visitor==0.1.3
wadllib==1.3.6
Werkzeug==2.2.3
WTForms==3.0.1
wxPython==4.0.7
xdg==5
xkit==0.0.0
yarg==0.1.9
zipp==1.0.0
